+++ servers/nanobanana/nanobanana_mcp_server/__main__.py
@@ -1,0 +1,4 @@
+"""Allow running as: python -m nanobanana_mcp_server"""
+from . import main
+
+main()